First changes for rust cli improvements
