+++ client/Conductor.Client.Cli/Cargo.toml
@@ -5,7 +5,8 @@ edition = "2024"
 
 [dependencies]
 clap = { version = "4.5.51", features = ["derive"] }
-reqwest = "0.12.24"
+reqwest = { version = "0.12.24", features = ["json"] }
 serde = { version = "1.0.209", features = ["derive"] }
 serde_json = "1.0.127"
 tokio = { version = "1.48.0", features = ["full"] }
+uuid = "1.18.1"
