+++ engine/src/Conductor.Engine.Api/Conductor.Engine.Api.csproj
@@ -5,7 +5,7 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.11"/>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="10.0.1"/>
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="9.0.6" />
     </ItemGroup>
 
     <ItemGroup>
